revert back to uint32 for unsigned int

+++ gbdxtools/ipe/util.py
@@ -33,7 +33,7 @@ IPE_TO_DTYPE = {
     "SHORT": "short",
     "UNSIGNED_SHORT": "ushort",
     "INTEGER": "int32",
-    "UNSIGNED_INTEGER": "uint8",
+    "UNSIGNED_INTEGER": "uint32",
     "LONG": "int64",
     "UNSIGNED_LONG": "uint64",
     "FLOAT": "float32",
